Change implementation to run native Android Code
Resolves problematic code saying initialize(new
MyGdxGame(actionResolverAndroid), config); .... cannot be applied
By introducing a Game class and a Screen which shows a Toast.
This example was made taking as reference the following tutorial:

https://carlorodriguez.github.io/blog/2014/10/05/android-platform-specific-code-with-libgdx/
