+++ android/AndroidManifest.xml
@@ -6,6 +6,8 @@
 
     <uses-sdk android:minSdkVersion="8" android:targetSdkVersion="22" />
 
+    <uses-permission android:name="android.permission.ACCESS_NETWORK_STATE"/>
+
     <application
         android:allowBackup="true"
         android:icon="@drawable/ic_launcher"
