+++ core/src/com/mygdx/game/MyGdxGame.java
@@ -1,12 +1,9 @@
 package com.mygdx.game;
 
-import com.badlogic.gdx.Gdx;
-import com.badlogic.gdx.Screen;
-import com.badlogic.gdx.graphics.GL20;
-import com.badlogic.gdx.graphics.Texture;
+import com.badlogic.gdx.Game;
 import com.badlogic.gdx.graphics.g2d.SpriteBatch;
 
-public class MyGdxGame implements Screen {
+public class MyGdxGame extends Game {
 
     ActionResolver actionResolver;
 
@@ -15,47 +12,21 @@ public class MyGdxGame implements Screen {
     }
 
 	SpriteBatch batch;
-	Texture img;
-
-	@Override
-	public void render (float delta) {
-		Gdx.gl.glClearColor(1, 0, 0, 1);
-		Gdx.gl.glClear(GL20.GL_COLOR_BUFFER_BIT);
-		batch.begin();
-		batch.draw(img, 0, 0);
-		batch.end();
-	}
 
     @Override
-    public void resize(int width, int height) {
-    }
-
-    @Override
-    public void show() {
-        int a;
+    public void create() {
         batch = new SpriteBatch();
-        img = new Texture("badlogic.jpg");
-
-        actionResolver.getNetInfo();
-
+        this.setScreen(new MyScreen(this));
     }
 
     @Override
-    public void hide() {
-
-    }
-
-    @Override
-    public void pause() {
-    }
-
-    @Override
-    public void resume() {
+    public void render() {
+        super.render();
     }
 
     @Override
     public void dispose() {
-
+        batch.dispose();
     }
 
 }
